# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Chatterbox TTS Extended Plus API
  description: Advanced Text-to-Speech and Voice Conversion API with chunking, retry logic, and Whisper validation
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: http://localhost:7860
    description: Local development server
...
paths:
...
  /api/v1/tts:
    post:
      summary: Text-to-Speech Generation
      description: Generate speech from text with optional voice cloning and advanced features
      operationId: generateTTS
      tags:
        - TTS
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TTSRequest'
            examples:
              basic:
                summary: Basic TTS
                value:
                  text: "Hello, this is a basic TTS example."
                  export_formats: ["wav"]
              with_voice:
                summary: TTS with Voice Cloning
                value:
                  text: "This will be spoken in the reference voice style."
                  reference_audio_filename: "speaker1/formal.wav"
                  exaggeration: 0.7
                  temperature: 0.8
                  export_formats: ["wav", "mp3"]
      responses:
        '200':
          description: TTS generation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TTSResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/vc:
    post:
      summary: Voice Conversion
      description: Convert voice characteristics of input audio to match a target voice
      operationId: convertVoice
      tags:
        - Voice Conversion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VCRequest'
            examples:
              local_files:
                summary: Local Files
                value:
                  input_audio_source: "my_recording.wav"
                  target_voice_source: "speaker1/formal.wav"
                  export_formats: ["wav"]
              with_urls:
                summary: With URL Input
                value:
                  input_audio_source: "https://example.com/speech.wav"
                  target_voice_source: "target_voices/celebrity.wav"
                  chunk_sec: 30
                  export_formats: ["wav", "mp3"]
      responses:
        '200':
          description: Voice conversion successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VCResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  schemas:
    TTSRequest:
      type: object
      required:
        - text
      properties:
        text:
          type: string
          description: Text to synthesize
          maxLength: 10000
          example: "Hello, this is a test of the TTS API."
        reference_audio_filename:
          type: string
          description: Reference audio file for voice cloning
          example: "speaker1/formal.wav"
        exaggeration:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          default: 0.5
          description: Voice expressiveness level
        temperature:
          type: number
          format: float
          minimum: 0.0
          maximum: 2.0
          default: 0.75
          description: Generation randomness
        seed:
          type: integer
          default: 0
          description: Random seed (0 for random)
        export_formats:
          type: array
          items:
            type: string
            enum: ["wav", "mp3", "flac"]
          default: ["wav", "mp3"]
          description: Output audio formats
        bypass_whisper_checking:
          type: boolean
          default: false
          description: Skip Whisper quality validation
        whisper_model_name:
          type: string
          enum: ["tiny", "base", "small", "medium", "large"]
          default: "medium"
          description: Whisper model size
    VCRequest:
      type: object
      required:
        - input_audio_source
        - target_voice_source
      properties:
        input_audio_source:
          type: string
          description: Source audio file or URL
          example: "my_recording.wav"
        target_voice_source:
          type: string
          description: Target voice file or URL
          example: "speaker1/formal.wav"
        chunk_sec:
          type: integer
          minimum: 10
          maximum: 300
          default: 60
          description: Chunk duration in seconds
        overlap_sec:
          type: number
          format: float
          minimum: 0.0
          maximum: 10.0
          default: 0.1
          description: Overlap for seamless joining
        export_formats:
          type: array
          items:
            type: string
            enum: ["wav", "mp3", "flac"]
          default: ["wav", "mp3"]
          description: Output audio formats
    AudioFile:
      type: object
      properties:
        format:
          type: string
          description: Audio file format
          example: "wav"
        filename:
          type: string
          description: Generated filename
          example: "tts_output_1234567890_42.wav"
        url:
          type: string
          description: HTTP URL for file access
          example: "/outputs/tts_output_1234567890_42.wav"
    BaseResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Operation success status
        message:
          type: string
          description: Response message
    TTSResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            output_files:
              type: array
              items:
                $ref: '#/components/schemas/AudioFile'
            generation_seed_used:
              type: integer
              description: Seed used for generation
            processing_time_seconds:
              type: number
              format: float
              description: Processing time
    VCResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            output_files:
              type: array
              items:
                $ref: '#/components/schemas/AudioFile'
            processing_time_seconds:
              type: number
              format: float
              description: Processing time
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Error message
        error_code:
          type: string
          description: Error code
        detail:
          type: string
          description: Technical details
  responses:
    BadRequest:
      description: Bad request - Invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'